修改卡路里为运动量；增加升级功能

+++ src/com/runrunfast/homegym/bean/Course.java
@@ -60,13 +60,13 @@ public class Course implements Serializable{
 	public static class GroupDetail implements Serializable{
 		public int count;
 		public int weight;
-		public float kcal;
+		public int kcal;
 		
 		public GroupDetail(){
 			
 		}
 		
-		public GroupDetail(int count, int weight, float kcal){
+		public GroupDetail(int count, int weight, int kcal){
 			this.count = count;
 			this.weight = weight;
 			this.kcal = kcal;
