+++ src/com/runrunfast/homegym/record/TotalRecord.java
@@ -1,8 +1,8 @@
 package com.runrunfast.homegym.record;
 
 public class TotalRecord {
-	public float total_kcal;
+	public int total_kcal;
 	public int total_time;
 	public int total_days;
-	public float total_food;
+	public int total_food;
 }
